Take account of the warnings categories to determine minimum version.
The Perl version that introduced each category was determined by
trolling through lib/warnings.pm in each Perl kit beginning with 5.6.0,
making note of when each version first appeared.

I considered using the table in regen/warnings.pl, but some of the
versions in that table seem off to me ('all' is present in 5.6, but the
table says 5.8), and it does not contain retracted warnings categories
(e.g. 'chmod').