Return '5.006' minimum if any 'use warnings' found.
Said less tersely: the previous version of the code returned nothing
unless some 'use warnings' had an argument. This version returns '5.006'
if 'use warnings' is found but no arguments are found. This makes no
difference in practice since v5.6 is the assumed minimum, but I'm
paranoid and it makes me feel better.